working on test 5
Former-commit-id: 1aba4334225e411df304d2cdb150fd7d5b2f02d6

+++ Sandbox/src/gbc/cpu.cpp
@@ -423,7 +423,7 @@ namespace GBC
 
     if(r == IMode::LOW)
     {
-      val = *src_value;
+      val = (uint8_t)*src_value;
     }
 
     if(r == IMode::ALL)
@@ -477,7 +477,7 @@ namespace GBC
       Set_Half_Carry(*dest_register, val, false);
       Set_Carry_Plus(*dest_register, val, false);
       *dest_register += val;
-    }
+   }
 
     return 0;
   }
